🔐 Phase 2: Add session token validation middleware (#51)
- Create src/session/manager.ts with comprehensive token management
- Create src/session/index.ts barrel export
- Integrate session module as single source of truth
- Remove ~460 lines of duplicate session code from index.ts
- Add dependency injection for tree/agent validation
- Add token refresh (heartbeat) capability
- Add periodic token cleanup
