+++ src/index.ts
@@ -18,6 +18,22 @@ import { EventBus } from './websocket/events.js';
 import { PinocchioWebSocket } from './websocket/server.js';
 import { WebSocketConfig, SpawnHandlerArgs, SpawnHandlerResult, TokenValidationResult as WsTokenValidationResult } from './websocket/types.js';
 import { ProgressTracker } from './websocket/progress.js';
+import {
+  initSessionManager,
+  generateSessionToken as generateToken,
+  storeSessionToken,
+  getSessionToken,
+  validateSessionToken,
+  invalidateSessionToken,
+  invalidateTokensForAgent,
+  invalidateTokensForTree,
+  saveTokenState as saveSessionTokenState,
+  loadTokenState as loadSessionTokenState,
+  startTokenCleanup,
+  stopTokenCleanup,
+  SessionToken,
+  TokenValidationResult,
+} from './session/index.js';
 
 // PR #32 FIX: Docker stream demultiplexer
 // Docker multiplexed streams have an 8-byte header per message:
@@ -176,7 +192,7 @@ const TREES_STATE_FILE = path.join(os.homedir(), ".config", "pinocchio", "trees.
 
 // Issue #47: Nested spawning configuration
 // Controls limits and behavior for agent-spawned agents
-interface NestedSpawnConfig {
+export interface NestedSpawnConfig {
   maxNestingDepth: number;       // Max depth of spawning (default: 2 = parent + child)
   maxAgentsPerTree: number;      // Max total agents in a spawn tree (default: 10)
   enableRecursiveSpawn: boolean; // Allow nested spawning at all (default: true)
@@ -431,7 +447,7 @@ async function rotateAuditLogIfNeeded(): Promise<void> {
 
 // AUDIT FIX #9: Append an audit event to the log file
 // Events are stored as JSONL (one JSON object per line)
-async function auditLog(event: string, data: Record<string, unknown>, agentId?: string): Promise<void> {
+export async function auditLog(event: string, data: Record<string, unknown>, agentId?: string): Promise<void> {
   try {
     // Ensure config directory exists
     const dir = path.dirname(AUDIT_LOG_FILE);
@@ -706,7 +722,7 @@ async function saveTreeState(): Promise<void> {
 const docker = new Docker();
 
 // Agent metadata for tracking
-interface AgentMetadata {
+export interface AgentMetadata {
   id: string;
   task: string;
   workspacePath: string;
@@ -734,7 +750,7 @@ const runningAgents = new Map<string, Docker.Container>();
 const agentMetadata = new Map<string, AgentMetadata>();
 
 // Issue #46: SpawnTree tracks an entire hierarchy of spawned agents as a single unit
-interface SpawnTree {
+export interface SpawnTree {
   treeId: string;                        // Unique tree identifier (matches AgentMetadata.treeId)
   rootAgentId: string;                   // ID of the root agent that started this tree
   totalAgents: number;                   // Count of all agents in this tree (root + children)
@@ -762,8 +778,13 @@ function createSpawnTree(rootAgentId: string, treeId: string): SpawnTree {
 }
 
 // Issue #46: Get a spawn tree by ID
-function getSpawnTree(treeId: string): SpawnTree | undefined {
+export function getSpawnTree(treeId: string): SpawnTree | undefined {
   return spawnTrees.get(treeId);
+}
+
+// Issue #51: Get agent metadata by ID (for session module)
+export function getAgentMetadata(agentId: string): AgentMetadata | undefined {
+  return agentMetadata.get(agentId);
 }
 
 // Issue #46: Update tree agent count (delta can be positive or negative)
@@ -828,262 +849,6 @@ function checkAndTerminateTree(treeId: string): void {
 
 // WebSocket server instance (initialized in main if enabled)
 let wsServer: PinocchioWebSocket | null = null;
-
-// =============================================================================
-// Issue #48: Session Token System for Nested Agent Authentication
-// =============================================================================
-// Session tokens are used to authenticate spawn requests from child agents.
-// Tokens are HMAC-signed to prevent forgery and contain hierarchy information
-// for validation. Tokens are short-lived and invalidated when parent terminates.
-
-// Session token configuration
-const SESSION_TOKEN_CONFIG = {
-  // Default token lifetime: 1 hour (capped by agent timeout)
-  defaultTokenLifetimeMs: 60 * 60 * 1000,
-  // Token is 32 bytes hex-encoded (64 characters)
-  tokenBytes: 32,
-  // HMAC algorithm for signing
-  hmacAlgorithm: 'sha256' as const,
-};
-
-// Server secret for HMAC signing (generated on startup)
-// In production, this should be loaded from secure storage
-// Regenerated on server restart, which invalidates all existing tokens
-const serverSecret = crypto.randomBytes(32);
-
-// Issue #48: SessionToken interface
-// Represents a cryptographically signed token for agent authentication
-interface SessionToken {
-  token: string;              // Cryptographically random, 32 bytes hex
-  signature: string;          // HMAC signature of token payload
-  agentId: string;            // Agent this token belongs to
-  treeId: string;             // Spawn tree for validation
-  parentAgentId?: string;     // Parent for hierarchy validation
-  depth: number;              // Current nesting depth
-  maxDepth: number;           // Configured max depth
-  issuedAt: number;           // Unix timestamp (ms)
-  expiresAt: number;          // Token expiry (ms)
-  permissions: {
-    canSpawn: boolean;        // Can this agent spawn children?
-    inheritGitHubToken: boolean;  // Can inherit parent's GitHub token?
-  };
-}
-
-// Issue #48: Serializable session token for persistence
-interface PersistedSessionToken {
-  token: string;
-  signature: string;
-  agentId: string;
-  treeId: string;
-  parentAgentId?: string;
-  depth: number;
-  maxDepth: number;
-  issuedAt: number;
-  expiresAt: number;
-  permissions: {
-    canSpawn: boolean;
-    inheritGitHubToken: boolean;
-  };
-}
-
-// Issue #48: In-memory token store
-// Map from token string to SessionToken for O(1) lookup
-const sessionTokens = new Map<string, SessionToken>();
-
-// Issue #48: State file path for token persistence
-const TOKENS_STATE_FILE = path.join(os.homedir(), ".config", "pinocchio", "tokens.json");
-
-// Issue #48: Generate the HMAC signature for a token payload
-// This prevents token forgery - only the server can create valid signatures
-function signTokenPayload(payload: {
-  token: string;
-  agentId: string;
-  treeId: string;
-  parentAgentId?: string;
-  depth: number;
-  maxDepth: number;
-  issuedAt: number;
-  expiresAt: number;
-  permissions: { canSpawn: boolean; inheritGitHubToken: boolean };
-}): string {
-  const hmac = crypto.createHmac(SESSION_TOKEN_CONFIG.hmacAlgorithm, serverSecret);
-  // Create deterministic string representation of payload
-  const dataToSign = JSON.stringify({
-    token: payload.token,
-    agentId: payload.agentId,
-    treeId: payload.treeId,
-    parentAgentId: payload.parentAgentId || '',
-    depth: payload.depth,
-    maxDepth: payload.maxDepth,
-    issuedAt: payload.issuedAt,
-    expiresAt: payload.expiresAt,
-    canSpawn: payload.permissions.canSpawn,
-    inheritGitHubToken: payload.permissions.inheritGitHubToken,
-  });
-  hmac.update(dataToSign);
-  return hmac.digest('hex');
-}
-
-// Issue #48: Verify that a token's signature is valid
-function verifyTokenSignature(token: SessionToken): boolean {
-  const expectedSignature = signTokenPayload({
-    token: token.token,
-    agentId: token.agentId,
-    treeId: token.treeId,
-    parentAgentId: token.parentAgentId,
-    depth: token.depth,
-    maxDepth: token.maxDepth,
-    issuedAt: token.issuedAt,
-    expiresAt: token.expiresAt,
-    permissions: token.permissions,
-  });
-  // Use timing-safe comparison to prevent timing attacks
-  try {
-    return crypto.timingSafeEqual(
-      Buffer.from(token.signature, 'hex'),
-      Buffer.from(expectedSignature, 'hex')
-    );
-  } catch {
-    // Length mismatch or invalid hex - signature is invalid
-    return false;
-  }
-}
-
-// Issue #48: Generate a new session token for an agent
-// Called when spawning a new agent (root or child)
-function generateSessionToken(
-  agentId: string,
-  treeId: string,
-  parentAgentId: string | undefined,
-  depth: number,
-  nestedSpawnConfig: NestedSpawnConfig,
-  timeoutMs: number,
-  inheritGitHubToken: boolean = false
-): SessionToken {
-  // Generate cryptographically random token
-  const token = crypto.randomBytes(SESSION_TOKEN_CONFIG.tokenBytes).toString('hex');
-
-  const now = Date.now();
-  // Token expiry is the minimum of:
-  // 1. Default token lifetime (1 hour)
-  // 2. Agent timeout (so token expires when agent should be done)
-  const expiresAt = now + Math.min(SESSION_TOKEN_CONFIG.defaultTokenLifetimeMs, timeoutMs);
-
-  // Determine spawn permission based on depth and config
-  const canSpawn = nestedSpawnConfig.enableRecursiveSpawn &&
-                   depth < nestedSpawnConfig.maxNestingDepth;
-
-  const sessionToken: SessionToken = {
-    token,
-    signature: '', // Will be set after signing
-    agentId,
-    treeId,
-    parentAgentId,
-    depth,
-    maxDepth: nestedSpawnConfig.maxNestingDepth,
-    issuedAt: now,
-    expiresAt,
-    permissions: {
-      canSpawn,
-      inheritGitHubToken,
-    },
-  };
-
-  // Sign the token
-  sessionToken.signature = signTokenPayload(sessionToken);
-
-  return sessionToken;
-}
-
-// Issue #48: Store a session token
-function storeSessionToken(sessionToken: SessionToken): void {
-  sessionTokens.set(sessionToken.token, sessionToken);
-  console.error(`[pinocchio] Stored session token for agent: ${sessionToken.agentId} (expires: ${new Date(sessionToken.expiresAt).toISOString()})`);
-
-  // Audit log token generation (don't log the actual token for security)
-  auditLog("token.generated", {
-    treeId: sessionToken.treeId,
-    depth: sessionToken.depth,
-    maxDepth: sessionToken.maxDepth,
-    canSpawn: sessionToken.permissions.canSpawn,
-    inheritGitHubToken: sessionToken.permissions.inheritGitHubToken,
-    expiresAt: new Date(sessionToken.expiresAt).toISOString(),
-  }, sessionToken.agentId).catch(() => {}); // Fire and forget
-}
-
-// Issue #48: Retrieve a session token by token string
-function getSessionToken(tokenString: string): SessionToken | undefined {
-  return sessionTokens.get(tokenString);
-}
-
-// Issue #48: Validation result for session tokens
-interface TokenValidationResult {
-  valid: boolean;
-  token?: SessionToken;
-  error?: string;
-}
-
-// Issue #48: Validate a session token
-// Checks: signature, expiry, tree active, parent running (if applicable)
-function validateSessionToken(tokenString: string): TokenValidationResult {
-  // Check token exists in store
-  const token = sessionTokens.get(tokenString);
-  if (!token) {
-    // Log potential token replay or forgery attempt
-    auditLog("token.validation_failed", {
-      reason: 'Token not found',
-      // Log a prefix of the token for debugging (not full token for security)
-      tokenPrefix: tokenString.slice(0, 8) + '...',
-    }).catch(() => {});
-    return { valid: false, error: 'Token not found' };
-  }
-
-  // Verify signature (protects against tampering)
-  if (!verifyTokenSignature(token)) {
-    console.error(`[pinocchio] Token signature verification failed for agent: ${token.agentId}`);
-    // This is a serious security event - potential tampering
-    auditLog("token.signature_invalid", {
-      reason: 'Signature verification failed',
-      treeId: token.treeId,
-      depth: token.depth,
-    }, token.agentId).catch(() => {});
-    return { valid: false, error: 'Invalid token signature' };
-  }
-
-  // Check expiry
-  const now = Date.now();
-  if (now > token.expiresAt) {
-    auditLog("token.expired", {
-      reason: 'Token expired',
-      expiredAt: new Date(token.expiresAt).toISOString(),
-      treeId: token.treeId,
-    }, token.agentId).catch(() => {});
-    return { valid: false, error: 'Token expired' };
-  }
-
-  // Check if the spawn tree is still active
-  const tree = getSpawnTree(token.treeId);
-  if (!tree) {
-    return { valid: false, error: 'Spawn tree not found' };
-  }
-  if (tree.status === 'terminated') {
-    return { valid: false, error: 'Spawn tree terminated' };
-  }
-
-  // If token has a parent, verify parent is still running
-  if (token.parentAgentId) {
-    const parentMetadata = agentMetadata.get(token.parentAgentId);
-    if (!parentMetadata) {
-      return { valid: false, error: 'Parent agent not found' };
-    }
-    if (parentMetadata.status !== 'running') {
-      return { valid: false, error: 'Parent agent no longer running' };
-    }
-  }
-
-  // All checks passed
-  return { valid: true, token };
-}
 
 // Issue #50: Wrapper for validateSessionToken for WebSocket server
 // Converts internal TokenValidationResult to the WebSocket module's format
@@ -1202,200 +967,6 @@ async function handleSpawnFromHttp(args: SpawnHandlerArgs): Promise<SpawnHandler
       success: false,
       error: message,
     };
-  }
-}
-
-// Issue #48: Invalidate a single session token
-function invalidateSessionToken(tokenString: string): boolean {
-  const token = sessionTokens.get(tokenString);
-  if (token) {
-    sessionTokens.delete(tokenString);
-    console.error(`[pinocchio] Invalidated session token for agent: ${token.agentId}`);
-    return true;
-  }
-  return false;
-}
-
-// Issue #48: Invalidate all tokens belonging to a specific agent
-function invalidateTokensForAgent(agentId: string): number {
-  let count = 0;
-  for (const [tokenString, token] of sessionTokens) {
-    if (token.agentId === agentId) {
-      sessionTokens.delete(tokenString);
-      count++;
-    }
-  }
-  if (count > 0) {
-    console.error(`[pinocchio] Invalidated ${count} token(s) for agent: ${agentId}`);
-  }
-  return count;
-}
-
-// Issue #48: Invalidate all tokens in a spawn tree
-// Called when a tree is terminated or when cleanup is needed
-function invalidateTokensForTree(treeId: string): number {
-  let count = 0;
-  for (const [tokenString, token] of sessionTokens) {
-    if (token.treeId === treeId) {
-      sessionTokens.delete(tokenString);
-      count++;
-    }
-  }
-  if (count > 0) {
-    console.error(`[pinocchio] Invalidated ${count} token(s) for tree: ${treeId}`);
-  }
-  return count;
-}
-
-// Issue #48: Clean up expired tokens (called periodically)
-function cleanupExpiredTokens(): number {
-  const now = Date.now();
-  let count = 0;
-  for (const [tokenString, token] of sessionTokens) {
-    if (now > token.expiresAt) {
-      sessionTokens.delete(tokenString);
-      count++;
-    }
-  }
-  if (count > 0) {
-    console.error(`[pinocchio] Cleaned up ${count} expired token(s)`);
-  }
-  return count;
-}
-
-// Issue #48: Save token state to disk
-async function saveTokenState(): Promise<void> {
-  try {
-    const dir = path.dirname(TOKENS_STATE_FILE);
-    await fs.mkdir(dir, { recursive: true, mode: 0o700 });
-
-    // Convert Map to array for serialization
-    const persisted: PersistedSessionToken[] = [];
-    const now = Date.now();
-
-    for (const [, token] of sessionTokens) {
-      // Only persist non-expired tokens
-      if (token.expiresAt > now) {
-        persisted.push({
-          token: token.token,
-          signature: token.signature,
-          agentId: token.agentId,
-          treeId: token.treeId,
-          parentAgentId: token.parentAgentId,
-          depth: token.depth,
-          maxDepth: token.maxDepth,
-          issuedAt: token.issuedAt,
-          expiresAt: token.expiresAt,
-          permissions: { ...token.permissions },
-        });
-      }
-    }
-
-    // Use atomic write pattern (temp file + rename)
-    const tempFile = `${TOKENS_STATE_FILE}.tmp.${process.pid}`;
-    await fs.writeFile(tempFile, JSON.stringify(persisted, null, 2), { mode: 0o600 });
-    await fs.chmod(tempFile, 0o600);
-    await fs.rename(tempFile, TOKENS_STATE_FILE);
-  } catch (error) {
-    console.error(`[pinocchio] Warning: Could not save token state: ${error}`);
-  }
-}
-
-// Issue #48: Load token state from disk on startup
-// Note: Tokens loaded from disk will have invalid signatures since serverSecret
-// is regenerated on restart. This is intentional for security - tokens from
-// previous server instances should not be trusted.
-async function loadTokenState(): Promise<void> {
-  try {
-    const data = await fs.readFile(TOKENS_STATE_FILE, "utf-8");
-    let persisted: PersistedSessionToken[];
-
-    try {
-      persisted = JSON.parse(data);
-    } catch (parseError) {
-      // JSON parse failed - back up corrupted file
-      const backupFile = `${TOKENS_STATE_FILE}.corrupted.${Date.now()}`;
-      console.error(`[pinocchio] Tokens state file corrupted, backing up to: ${backupFile}`);
-      try {
-        await fs.copyFile(TOKENS_STATE_FILE, backupFile);
-        await fs.chmod(backupFile, 0o600);
-      } catch (backupError) {
-        console.error(`[pinocchio] Warning: Could not create backup: ${backupError}`);
-      }
-      console.error(`[pinocchio] Starting with fresh token state due to parse error: ${parseError}`);
-      return;
-    }
-
-    const now = Date.now();
-    let loadedCount = 0;
-    let expiredCount = 0;
-    let invalidSignatureCount = 0;
-
-    for (const item of persisted) {
-      // Skip expired tokens
-      if (item.expiresAt <= now) {
-        expiredCount++;
-        continue;
-      }
-
-      const token: SessionToken = {
-        token: item.token,
-        signature: item.signature,
-        agentId: item.agentId,
-        treeId: item.treeId,
-        parentAgentId: item.parentAgentId,
-        depth: item.depth,
-        maxDepth: item.maxDepth,
-        issuedAt: item.issuedAt,
-        expiresAt: item.expiresAt,
-        permissions: { ...item.permissions },
-      };
-
-      // Verify signature - since serverSecret is regenerated on restart,
-      // all loaded tokens will have invalid signatures. We regenerate
-      // signatures for tokens that belong to still-running agents.
-      // This maintains security while preserving session continuity.
-      const agentMeta = agentMetadata.get(token.agentId);
-      if (agentMeta && agentMeta.status === 'running') {
-        // Re-sign the token with new server secret
-        token.signature = signTokenPayload(token);
-        sessionTokens.set(token.token, token);
-        loadedCount++;
-      } else {
-        // Agent not running, don't restore token
-        invalidSignatureCount++;
-      }
-    }
-
-    if (loadedCount > 0 || expiredCount > 0 || invalidSignatureCount > 0) {
-      console.error(`[pinocchio] Token state loaded: ${loadedCount} restored, ${expiredCount} expired, ${invalidSignatureCount} for non-running agents`);
-    }
-  } catch (error) {
-    if ((error as NodeJS.ErrnoException).code !== "ENOENT") {
-      console.error(`[pinocchio] Warning: Could not load token state: ${error}`);
-    }
-  }
-}
-
-// Issue #48: Start periodic token cleanup (every 5 minutes)
-let tokenCleanupInterval: NodeJS.Timeout | null = null;
-
-function startTokenCleanup(): void {
-  if (tokenCleanupInterval) {
-    return; // Already running
-  }
-  tokenCleanupInterval = setInterval(() => {
-    cleanupExpiredTokens();
-  }, 5 * 60 * 1000); // Every 5 minutes
-
-  // Don't prevent Node from exiting
-  tokenCleanupInterval.unref();
-}
-
-function stopTokenCleanup(): void {
-  if (tokenCleanupInterval) {
-    clearInterval(tokenCleanupInterval);
-    tokenCleanupInterval = null;
   }
 }
 
@@ -2501,7 +2072,7 @@ async function spawnDockerAgent(args: {
     // Issue #48: Generate session token for this agent
     // Root agents get tokens that allow them to spawn children (if config permits)
     // Issue #49: Child agents also get tokens, with inherited hierarchy info
-    const agentSessionToken = generateSessionToken(
+    const agentSessionToken = generateToken(
       agentId,
       treeId,
       args.parent_agent_id,  // parentAgentId (undefined for root agents)
@@ -2558,7 +2129,7 @@ async function spawnDockerAgent(args: {
     await saveTreeState();
 
     // Issue #48: Persist token state when token is created
-    await saveTokenState();
+    await saveSessionTokenState();
 
     // Start the container
     await container.start();
@@ -2741,7 +2312,7 @@ async function spawnDockerAgent(args: {
     await saveTreeState();
 
     // Issue #48: Persist token state after invalidation
-    await saveTokenState();
+    await saveSessionTokenState();
 
     // Clean up container (keep metadata for status queries)
     try {
@@ -2811,7 +2382,7 @@ async function spawnDockerAgent(args: {
       await saveTreeState();
 
       // Issue #48: Persist token state after invalidation
-      await saveTokenState();
+      await saveSessionTokenState();
     }
 
     // SECURITY FIX #8: Clean up token file on error
@@ -2972,7 +2543,7 @@ async function monitorAgent(agentId: string, container: Docker.Container, timeou
     await saveTreeState();
 
     // Issue #48: Persist token state after invalidation
-    await saveTokenState();
+    await saveSessionTokenState();
 
     // SECURITY FIX #8: Clean up token file after container completes
     if (metadata.tokenFilePath) {
@@ -3027,7 +2598,7 @@ async function monitorAgent(agentId: string, container: Docker.Container, timeou
     await saveTreeState();
 
     // Issue #48: Persist token state after invalidation
-    await saveTokenState();
+    await saveSessionTokenState();
 
     // SECURITY FIX #8: Clean up token file even on error
     if (metadata.tokenFilePath) {
@@ -3247,7 +2818,7 @@ async function stopDockerAgent(args: { agent_id: string }) {
       await saveTreeState();
 
       // Issue #48: Persist token state after invalidation
-      await saveTokenState();
+      await saveSessionTokenState();
     }
 
     // AUDIT FIX #9: Log agent stop event
@@ -3794,9 +3365,16 @@ async function main() {
   // Issue #46: Load persisted spawn tree state on startup
   await loadTreeState();
 
+  // Issue #51: Initialize session manager with dependencies from index.ts
+  initSessionManager({
+    getSpawnTree: (treeId) => getSpawnTree(treeId),
+    getAgentMetadata: (agentId) => getAgentMetadata(agentId),
+    auditLog,
+  });
+
   // Issue #48: Load persisted session token state on startup
   // Note: Tokens are re-signed with new server secret for running agents
-  await loadTokenState();
+  await loadSessionTokenState();
 
   // Issue #48: Start periodic token cleanup
   startTokenCleanup();
